refactor: split simulator responsibilities
