+++ internal/sim/simulator.go
@@ -2,7 +2,6 @@
 package sim
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"math"
@@ -250,428 +249,6 @@ func NewSimulator(clusterID string, cfg *config.SimulationConfig, writer Telemet
 	return sim
 }
 
-// Run starts the simulation loop and stops when the context is done.
-func (s *Simulator) Run(ctx context.Context) {
-	log.Printf("[Simulator] starting with tick interval %s", s.tickInterval)
-	ticker := time.NewTicker(s.tickInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ticker.C:
-			s.tick()
-		case <-ctx.Done():
-			log.Println("[Simulator] stopping...")
-			return
-		}
-	}
-}
-
-// tick generates telemetry and writes it.
-func (s *Simulator) tick() {
-	var batch []telemetry.TelemetryRow
-	var detections []enemy.DetectionRow
-
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	s.messagesSent = 0
-
-	var allDrones []*telemetry.Drone
-	for _, f := range s.fleets {
-		allDrones = append(allDrones, f.Drones...)
-	}
-	if s.enemyEng != nil {
-		for _, en := range s.enemyEng.Enemies {
-			s.enemyPrevPositions[en.ID] = en.Position
-		}
-		s.enemyEng.Step(allDrones)
-	}
-
-	for _, fleet := range s.fleets {
-		for _, drone := range fleet.Drones {
-			row, ok := s.updateDrone(drone)
-			if !ok {
-				continue
-			}
-			if s.chaosMode {
-				s.injectChaos(drone, &row)
-			}
-			batch = append(batch, row)
-			detections = append(detections, s.processDetections(&fleet, drone)...)
-		}
-	}
-
-	s.reassignFollowers()
-
-	// Batch support if writer implements WriteBatch
-	if bw, ok := s.writer.(batchWriter); ok {
-		if err := bw.WriteBatch(batch); err != nil {
-			log.Printf("[Simulator] batch write failed: %v", err)
-		}
-	} else {
-		for _, row := range batch {
-			if err := s.writer.Write(row); err != nil {
-				log.Printf("[Simulator] write failed for drone %s: %v", row.DroneID, err)
-			}
-		}
-	}
-
-	// Write enemy detections if any
-	if len(detections) > 0 && s.detectionWriter != nil {
-		if bw, ok := s.detectionWriter.(batchDetectionWriter); ok {
-			if err := bw.WriteDetections(detections); err != nil {
-				log.Printf("[Simulator] detection batch write failed: %v", err)
-			}
-		} else {
-			for _, d := range detections {
-				if err := s.detectionWriter.WriteDetection(d); err != nil {
-					log.Printf("[Simulator] detection write failed: %v", err)
-				}
-			}
-		}
-	}
-}
-
-func (s *Simulator) updateDrone(drone *telemetry.Drone) (telemetry.TelemetryRow, bool) {
-	if drone.FollowTarget != nil && (s.rand.Float64() < s.commLoss || drone.Status == telemetry.StatusFailure) {
-		s.removeAssignment(drone)
-	}
-	row := s.teleGen.GenerateTelemetry(drone)
-	if s.rand.Float64() < drone.SensorErrorRate {
-		row.Lat += s.rand.Float64()*sensorErrorMaxOffset*2 - sensorErrorMaxOffset
-		row.Lon += s.rand.Float64()*sensorErrorMaxOffset*2 - sensorErrorMaxOffset
-	}
-	if s.rand.Float64() < drone.BatteryAnomalyRate {
-		drop := s.rand.Float64()*20 + 10
-		drone.Battery -= drop
-		if drone.Battery < 0 {
-			drone.Battery = 0
-		}
-		row.Battery = drone.Battery
-	}
-	if s.rand.Float64() < drone.DropoutRate {
-		return telemetry.TelemetryRow{}, false
-	}
-	return row, true
-}
-
-func (s *Simulator) injectChaos(drone *telemetry.Drone, row *telemetry.TelemetryRow) {
-	if s.rand.Float64() < 0.1 {
-		row.Status = telemetry.StatusFailure
-		drone.Status = telemetry.StatusFailure
-	}
-	extra := s.rand.Float64() * 5
-	drone.Battery -= extra
-	if drone.Battery < 0 {
-		drone.Battery = 0
-	}
-	row.Battery = drone.Battery
-}
-
-func (s *Simulator) processDetections(fleet *DroneFleet, drone *telemetry.Drone) []enemy.DetectionRow {
-	if s.enemyEng == nil {
-		return nil
-	}
-	var detections []enemy.DetectionRow
-	for _, en := range s.enemyEng.Enemies {
-		dist := distanceMeters(drone.Position.Lat, drone.Position.Lon, en.Position.Lat, en.Position.Lon)
-		if dist > s.detectionRadiusM {
-			continue
-		}
-		conf := 100 * (1 - dist/s.detectionRadiusM)
-		conf *= 1 - s.terrainOcclusion
-		conf *= 1 - s.weatherImpact
-		if s.sensorNoise > 0 {
-			conf += s.rand.NormFloat64() * s.sensorNoise * conf
-		}
-		if conf < 0 {
-			conf = 0
-		} else if conf > 100 {
-			conf = 100
-		}
-		d := enemy.DetectionRow{
-			ClusterID:  s.clusterID,
-			DroneID:    drone.ID,
-			EnemyID:    en.ID,
-			EnemyType:  en.Type,
-			Lat:        en.Position.Lat,
-			Lon:        en.Position.Lon,
-			Alt:        en.Position.Alt,
-			Confidence: conf,
-			Timestamp:  s.now().UTC(),
-		}
-		detections = append(detections, d)
-		if conf >= s.followConfidence {
-			s.assignFollower(fleet, drone, en, conf)
-		}
-	}
-	return detections
-}
-
-func (s *Simulator) sendCommand() bool {
-	if s.bandwidthLimit > 0 && s.messagesSent >= s.bandwidthLimit {
-		return false
-	}
-	s.messagesSent++
-	if s.rand.Float64() < s.commLoss {
-		return false
-	}
-	return true
-}
-
-func (s *Simulator) removeAssignment(drone *telemetry.Drone) {
-	enemyID, ok := s.droneAssignments[drone.ID]
-	if ok {
-		followers := s.enemyFollowers[enemyID]
-		nf := followers[:0]
-		for _, id := range followers {
-			if id != drone.ID {
-				nf = append(nf, id)
-			}
-		}
-		if len(nf) == 0 {
-			delete(s.enemyFollowers, enemyID)
-			delete(s.enemyFollowerTargets, enemyID)
-		} else {
-			s.enemyFollowers[enemyID] = nf
-		}
-		delete(s.droneAssignments, drone.ID)
-	}
-	drone.FollowTarget = nil
-}
-
-func (s *Simulator) selectReplacement() *telemetry.Drone {
-	var best *telemetry.Drone
-	for _, f := range s.fleets {
-		for _, d := range f.Drones {
-			if d.Status != telemetry.StatusOK || d.FollowTarget != nil {
-				continue
-			}
-			if _, assigned := s.droneAssignments[d.ID]; assigned {
-				continue
-			}
-			if best == nil || d.Battery > best.Battery {
-				best = d
-			}
-		}
-	}
-	return best
-}
-
-// cleanupFollowers removes invalid followers for an enemy and returns remaining active IDs.
-func (s *Simulator) cleanupFollowers(enemyID string, followers []string) []string {
-	active := followers[:0]
-	for _, id := range followers {
-		d := s.droneIndex[id]
-		if d != nil && d.FollowTarget != nil && d.Status == telemetry.StatusOK {
-			active = append(active, id)
-			continue
-		}
-		delete(s.droneAssignments, id)
-		if d != nil {
-			d.FollowTarget = nil
-		}
-	}
-	return active
-}
-
-// selectCandidates finds up to missing replacement drones and reserves their assignments.
-func (s *Simulator) selectCandidates(missing int) []*telemetry.Drone {
-	var cands []*telemetry.Drone
-	for missing > 0 {
-		cand := s.selectReplacement()
-		if cand == nil || !s.sendCommand() {
-			break
-		}
-		s.droneAssignments[cand.ID] = "" // reserve to avoid reselection
-		cands = append(cands, cand)
-		missing--
-	}
-	return cands
-}
-
-// filterSendable reserves and returns drones that can receive a command.
-func (s *Simulator) filterSendable(cands []*telemetry.Drone) []*telemetry.Drone {
-	var selected []*telemetry.Drone
-	for _, c := range cands {
-		if s.sendCommand() {
-			s.droneAssignments[c.ID] = "" // reserve
-			selected = append(selected, c)
-		}
-	}
-	return selected
-}
-
-// applyAssignments finalizes assignments of candidates to an enemy.
-func (s *Simulator) applyAssignments(enemyID string, en *enemy.Enemy, cands []*telemetry.Drone) {
-	if len(cands) == 0 {
-		return
-	}
-	pts := s.interceptPoints(en, len(cands))
-	for i, d := range cands {
-		cp := pts[i]
-		d.FollowTarget = &cp
-		s.enemyFollowers[enemyID] = append(s.enemyFollowers[enemyID], d.ID)
-		s.droneAssignments[d.ID] = enemyID
-	}
-}
-
-func (s *Simulator) reassignFollowers() {
-	for enemyID, followers := range s.enemyFollowers {
-		active := s.cleanupFollowers(enemyID, followers)
-		desired := s.enemyFollowerTargets[enemyID]
-		missing := desired - len(active)
-		if missing <= 0 {
-			if len(active) == 0 {
-				delete(s.enemyFollowers, enemyID)
-				delete(s.enemyFollowerTargets, enemyID)
-			} else {
-				s.enemyFollowers[enemyID] = active
-			}
-			continue
-		}
-		s.enemyFollowers[enemyID] = active
-		en := s.enemyObjects[enemyID]
-		cands := s.selectCandidates(missing)
-		s.applyAssignments(enemyID, en, cands)
-		if len(s.enemyFollowers[enemyID]) == 0 {
-			delete(s.enemyFollowers, enemyID)
-			delete(s.enemyFollowerTargets, enemyID)
-		}
-	}
-}
-
-func (s *Simulator) assignFollower(fleet *DroneFleet, detecting *telemetry.Drone, en *enemy.Enemy, conf float64) {
-	s.enemyObjects[en.ID] = en
-	count, ok := s.swarmResponses[detecting.MovementPattern]
-	if !ok {
-		count = 0
-	}
-	if count >= 0 {
-		if conf > 90 {
-			count++
-		}
-		switch en.Type {
-		case enemy.EnemyVehicle, enemy.EnemyDrone:
-			count++
-		case enemy.EnemyDecoy:
-			if count > 0 {
-				count--
-			}
-		}
-		count += s.missionCriticality
-	}
-	if count == 0 {
-		cands := s.filterSendable([]*telemetry.Drone{detecting})
-		s.applyAssignments(en.ID, en, cands)
-		if len(cands) > 0 {
-			s.rebalanceFormation(fleet)
-		}
-		s.enemyFollowerTargets[en.ID] = len(s.enemyFollowers[en.ID])
-		return
-	}
-	if count < 0 {
-		var unassigned []*telemetry.Drone
-		for _, d := range fleet.Drones {
-			if d.FollowTarget == nil {
-				unassigned = append(unassigned, d)
-			}
-		}
-		selected := s.filterSendable(unassigned)
-		s.applyAssignments(en.ID, en, selected)
-		if len(selected) > 0 {
-			s.rebalanceFormation(fleet)
-		}
-		s.enemyFollowerTargets[en.ID] = len(s.enemyFollowers[en.ID])
-		return
-	}
-	var followers []*telemetry.Drone
-	for _, d := range fleet.Drones {
-		if d == detecting {
-			continue
-		}
-		if d.FollowTarget == nil {
-			followers = append(followers, d)
-			if len(followers) >= count {
-				break
-			}
-		}
-	}
-	if len(followers) == 0 {
-		cands := s.filterSendable([]*telemetry.Drone{detecting})
-		s.applyAssignments(en.ID, en, cands)
-		if len(cands) > 0 {
-			s.rebalanceFormation(fleet)
-		}
-		s.enemyFollowerTargets[en.ID] = len(s.enemyFollowers[en.ID])
-		return
-	}
-	selected := s.filterSendable(followers)
-	s.applyAssignments(en.ID, en, selected)
-	if len(selected) > 0 {
-		s.rebalanceFormation(fleet)
-	}
-	s.enemyFollowerTargets[en.ID] = len(s.enemyFollowers[en.ID])
-}
-
-func (s *Simulator) interceptPoints(en *enemy.Enemy, n int) []telemetry.Position {
-	points := make([]telemetry.Position, n)
-	target := en.Position
-	prev, ok := s.enemyPrevPositions[en.ID]
-	velLat := 0.0
-	velLon := 0.0
-	if ok {
-		velLat = target.Lat - prev.Lat
-		velLon = target.Lon - prev.Lon
-	}
-	predicted := telemetry.Position{Lat: target.Lat + velLat*5, Lon: target.Lon + velLon*5, Alt: target.Alt}
-	if n == 1 {
-		points[0] = predicted
-		return points
-	}
-	norm := math.Hypot(velLat, velLon)
-	var perpLat, perpLon float64
-	if norm != 0 {
-		perpLat = -velLon / norm
-		perpLon = velLat / norm
-	}
-	lateral := interceptLateralMeters
-	latStep := lateral / 111000
-	lonStep := lateral / (111000 * math.Cos(predicted.Lat*math.Pi/180))
-	for i := 0; i < n; i++ {
-		offset := float64(i) - float64(n-1)/2
-		points[i] = telemetry.Position{
-			Lat: predicted.Lat + offset*perpLat*latStep,
-			Lon: predicted.Lon + offset*perpLon*lonStep,
-			Alt: predicted.Alt,
-		}
-	}
-	return points
-}
-
-func (s *Simulator) rebalanceFormation(fleet *DroneFleet) {
-	var remaining []*telemetry.Drone
-	for _, d := range fleet.Drones {
-		if d.FollowTarget == nil {
-			remaining = append(remaining, d)
-		}
-	}
-	n := len(remaining)
-	if n == 0 {
-		return
-	}
-	region := remaining[0].HomeRegion
-	radius := region.RadiusKM * 1000 * 0.5
-	for i, d := range remaining {
-		angle := float64(i) / float64(n) * 2 * math.Pi
-		deltaLat := (radius * math.Cos(angle)) / 111000
-		deltaLon := (radius * math.Sin(angle)) / (111000 * math.Cos(region.CenterLat*math.Pi/180))
-		d.HomeRegion.CenterLat = region.CenterLat + deltaLat
-		d.HomeRegion.CenterLon = region.CenterLon + deltaLon
-	}
-}
-
 // ToggleChaos flips chaos mode on or off and returns the new state.
 func (s *Simulator) ToggleChaos() bool {
 	s.mu.Lock()
@@ -814,52 +391,6 @@ func (s *Simulator) MapSnapshot() MapData {
 		}
 	}
 	return MapData{Drones: drones, Enemies: enemies, Missions: missions}
-}
-
-// ObserverEvents returns a copy of all recorded mission events.
-func (s *Simulator) ObserverEvents() []ObserverEvent {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	events := make([]ObserverEvent, len(s.observerEvents))
-	copy(events, s.observerEvents)
-	return events
-}
-
-// ObserverStep sets the current event index and returns the event at that position.
-func (s *Simulator) ObserverStep(idx int) (ObserverEvent, bool) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if idx < 0 || idx >= len(s.observerEvents) {
-		return ObserverEvent{}, false
-	}
-	s.observerIdx = idx
-	return s.observerEvents[idx], true
-}
-
-// ObserverSetPerspective selects a drone to observe.
-func (s *Simulator) ObserverSetPerspective(id string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.observerPerspective = id
-	s.logObserverEvent("perspective", id)
-}
-
-// ObserverPerspective returns the current drone perspective.
-func (s *Simulator) ObserverPerspective() string {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.observerPerspective
-}
-
-// ObserverInjectCommand records a scripted command.
-func (s *Simulator) ObserverInjectCommand(cmd string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.logObserverEvent("command", cmd)
-}
-
-func (s *Simulator) logObserverEvent(t, details string) {
-	s.observerEvents = append(s.observerEvents, ObserverEvent{Timestamp: s.now().UTC(), Type: t, Details: details})
 }
 
 func generateDroneID(fleetName string, index int) string {
